+++ lib/ayadn/errors.rb
@@ -11,7 +11,7 @@ module Ayadn
       Logs.rec.debug "STACK: #{caller}"
       Logs.rec.error "--END--"
       puts "\n(error logged in #{Settings.config[:paths][:log]}/ayadn.log)\n".color(:blue)
-      Debug.err
+      Debug.err error
       exit
     end
     def self.error(status)